added buttons
displays six pokemon at a time now. hit next to show the next six, and
previous to show the previous six.
also added search by name, but it's VERY slow if the pokemon's dex
number is in the hundreds or worse, if it doesn't exist.
and input custom range, which lets you display a custom range of
pokemon.